Rename currentUrl hook module to useCurrentUrl
Now all hooks are in modules formatted as "useXyz".
