+++ src/hooks/useCurrentUrl.js
@@ -1,12 +1,13 @@
-/* Hook to get the current URL */
-
 import getConfig from 'next/config';
 import { useRouter } from 'next/router';
 
 const { publicRuntimeConfig } = getConfig();
 
-export const useCurrentUrl = () => {
+/* Hook to get the current URL. */
+function useCurrentUrl() {
   const { baseUrl } = publicRuntimeConfig;
   const router = useRouter();
   return new URL(router.pathname, baseUrl);
-};
+}
+
+export default useCurrentUrl;
